+++ tests/test_validator.py
@@ -134,6 +134,55 @@ def create_test_bag(src, dst, baginfo=None, algorithms=None) -> Bag:
             },
             False,
         ),
+        (
+            {
+                "Tag-Files-Required": ["meta/dir/"],
+                "Tag-Files-Allowed": ["meta/dir/*"],
+            },
+            True,
+        ),
+        (
+            {
+                "Tag-Files-Required": ["meta/dir/"],
+                "Tag-Files-Allowed": ["meta/dir/**"],
+            },
+            True,
+        ),
+        (
+            {
+                "Tag-Files-Required": ["meta/dir/"],
+                "Tag-Files-Allowed": ["meta/dir/a/*"],
+            },
+            True,
+        ),
+        (
+            {
+                "Tag-Files-Required": ["meta/dir/0"],
+                "Tag-Files-Allowed": ["meta/dir/[0-9]"],
+            },
+            True,
+        ),
+        (
+            {
+                "Tag-Files-Required": ["meta/dir/a"],
+                "Tag-Files-Allowed": ["meta/dir/[0-9]"],
+            },
+            False,
+        ),
+        (
+            {
+                "Tag-Files-Required": ["meta/a"],
+                "Tag-Files-Allowed": ["meta/*/*"],
+            },
+            False,
+        ),
+        (
+            {
+                "Tag-Files-Required": ["meta/a/b"],
+                "Tag-Files-Allowed": ["meta/*/*"],
+            },
+            True,
+        ),
         ({"Payload-Files-Required": None}, False),
         ({"Payload-Files-Allowed": None}, False),
         ({"Payload-Files-Required": []}, True),
@@ -163,9 +212,65 @@ def create_test_bag(src, dst, baginfo=None, algorithms=None) -> Bag:
         (
             {
                 "Payload-Files-Required": ["any/file"],
-                "Payload-Files-Allowed": ["meta/*"],
+                "Payload-Files-Allowed": ["data/*"],
             },
             False,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/dir/"],
+                "Payload-Files-Allowed": ["data/dir/*"],
+            },
+            True,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/dir/"],
+                "Payload-Files-Allowed": ["data/dir/**"],
+            },
+            True,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/dir/"],
+                "Payload-Files-Allowed": ["data/dir/a/*"],
+            },
+            True,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/dir/"],
+                "Payload-Files-Allowed": ["data/dir/[0-9]"],
+            },
+            True,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/dir/0"],
+                "Payload-Files-Allowed": ["data/dir/[0-9]"],
+            },
+            True,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/dir/a"],
+                "Payload-Files-Allowed": ["data/dir/[0-9]"],
+            },
+            False,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/a"],
+                "Payload-Files-Allowed": ["data/*/*"],
+            },
+            False,
+        ),
+        (
+            {
+                "Payload-Files-Required": ["data/a/b"],
+                "Payload-Files-Allowed": ["data/*/*"],
+            },
+            True,
         ),
     ],
 )